requirements.txt
asgiref==3.5.0
autopep8==1.6.0
awscli==1.22.47
boto3==1.20.46
botocore==1.23.47
colorama==0.4.3
Django==4.0.2
django-environ==0.8.1
djangorestframework==3.13.1
docutils==0.15.2
jmespath==0.10.0
pyasn1==0.4.8
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
